+++ server-data/resources/[esx_addons]/esx_addoninventory/server/main.lua
@@ -76,6 +76,7 @@ MySQL.ready(function()
 
 			local addonInventory    = CreateAddonInventory(name, nil, items)
 			SharedInventories[name] = addonInventory
+			GlobalState.SharedInventories = SharedInventories
 		end
 	end
 end)
@@ -90,6 +91,20 @@ end
 
 function GetSharedInventory(name)
 	return SharedInventories[name]
+end
+
+function AddSharedInventory(society)
+    if type(society) ~= 'table' or not society?.name or not society?.label then 
+		return
+	end
+    -- society (array) containing name (string) and label (string)
+    -- addon inventory:
+    MySQL.Async.execute('INSERT INTO addon_inventory (name, label, shared) VALUES (@name, @label, @shared)', {
+        ['name'] = society.name,
+        ['label'] = society.label,
+        ['shared'] = 1
+    })
+    SharedInventories[society.name] = CreateAddonInventory(society.name, nil, {})
 end
 
 AddEventHandler('esx_addoninventory:getInventory', function(name, owner, cb)